JAMES-1635 Modularise mailbox tables creation
