+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/modules/CassandraQuotaModule.java
@@ -1,0 +1,82 @@
+/****************************************************************
+ * Licensed to the Apache Software Foundation (ASF) under one   *
+ * or more contributor license agreements.  See the NOTICE file *
+ * distributed with this work for additional information        *
+ * regarding copyright ownership.  The ASF licenses this file   *
+ * to you under the Apache License, Version 2.0 (the            *
+ * "License"); you may not use this file except in compliance   *
+ * with the License.  You may obtain a copy of the License at   *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
+ * Unless required by applicable law or agreed to in writing,   *
+ * software distributed under the License is distributed on an  *
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
+ * KIND, either express or implied.  See the License for the    *
+ * specific language governing permissions and limitations      *
+ * under the License.                                           *
+ ****************************************************************/
+
+package org.apache.james.mailbox.cassandra.modules;
+
+import com.datastax.driver.core.schemabuilder.SchemaBuilder;
+import org.apache.james.backends.cassandra.components.CassandraIndex;
+import org.apache.james.backends.cassandra.components.CassandraModule;
+import org.apache.james.backends.cassandra.components.CassandraTable;
+import org.apache.james.backends.cassandra.components.CassandraType;
+import org.apache.james.mailbox.cassandra.table.CassandraCurrentQuota;
+import org.apache.james.mailbox.cassandra.table.CassandraDefaultMaxQuota;
+import org.apache.james.mailbox.cassandra.table.CassandraMaxQuota;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
+import static com.datastax.driver.core.DataType.bigint;
+import static com.datastax.driver.core.DataType.counter;
+import static com.datastax.driver.core.DataType.text;
+
+public class CassandraQuotaModule implements CassandraModule {
+
+    private final List<CassandraTable> tables;
+    private final List<CassandraIndex> index;
+    private final List<CassandraType> types;
+
+    public CassandraQuotaModule() {
+        tables = Arrays.asList(
+            new CassandraTable(CassandraCurrentQuota.TABLE_NAME,
+                    SchemaBuilder.createTable(CassandraCurrentQuota.TABLE_NAME)
+                    .ifNotExists()
+                    .addPartitionKey(CassandraCurrentQuota.QUOTA_ROOT, text())
+                    .addColumn(CassandraCurrentQuota.MESSAGE_COUNT, counter())
+                    .addColumn(CassandraCurrentQuota.STORAGE, counter())),
+            new CassandraTable(CassandraMaxQuota.TABLE_NAME,
+                    SchemaBuilder.createTable(CassandraMaxQuota.TABLE_NAME)
+                    .ifNotExists()
+                    .addPartitionKey(CassandraMaxQuota.QUOTA_ROOT, text())
+                    .addColumn(CassandraMaxQuota.MESSAGE_COUNT, bigint())
+                    .addColumn(CassandraMaxQuota.STORAGE, bigint())),
+            new CassandraTable(CassandraDefaultMaxQuota.TABLE_NAME,
+                    SchemaBuilder.createTable(CassandraDefaultMaxQuota.TABLE_NAME)
+                    .ifNotExists()
+                    .addPartitionKey(CassandraDefaultMaxQuota.TYPE, text())
+                    .addColumn(CassandraDefaultMaxQuota.VALUE, bigint())));
+        index = Collections.emptyList();
+        types = Collections.emptyList();
+    }
+
+    @Override
+    public List<CassandraTable> moduleTables() {
+        return tables;
+    }
+
+    @Override
+    public List<CassandraIndex> moduleIndex() {
+        return index;
+    }
+
+    @Override
+    public List<CassandraType> moduleTypes() {
+        return types;
+    }
+}
