+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/modules/CassandraUidAndModSeqModule.java
@@ -1,0 +1,73 @@
+/****************************************************************
+ * Licensed to the Apache Software Foundation (ASF) under one   *
+ * or more contributor license agreements.  See the NOTICE file *
+ * distributed with this work for additional information        *
+ * regarding copyright ownership.  The ASF licenses this file   *
+ * to you under the Apache License, Version 2.0 (the            *
+ * "License"); you may not use this file except in compliance   *
+ * with the License.  You may obtain a copy of the License at   *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
+ * Unless required by applicable law or agreed to in writing,   *
+ * software distributed under the License is distributed on an  *
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
+ * KIND, either express or implied.  See the License for the    *
+ * specific language governing permissions and limitations      *
+ * under the License.                                           *
+ ****************************************************************/
+
+package org.apache.james.mailbox.cassandra.modules;
+
+import com.datastax.driver.core.schemabuilder.SchemaBuilder;
+import org.apache.james.backends.cassandra.components.CassandraIndex;
+import org.apache.james.backends.cassandra.components.CassandraModule;
+import org.apache.james.backends.cassandra.components.CassandraTable;
+import org.apache.james.backends.cassandra.components.CassandraType;
+import org.apache.james.mailbox.cassandra.table.CassandraMessageModseqTable;
+import org.apache.james.mailbox.cassandra.table.CassandraMessageUidTable;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
+import static com.datastax.driver.core.DataType.bigint;
+import static com.datastax.driver.core.DataType.timeuuid;
+
+public class CassandraUidAndModSeqModule implements CassandraModule {
+
+    private final List<CassandraTable> tables;
+    private final List<CassandraIndex> index;
+    private final List<CassandraType> types;
+
+    public CassandraUidAndModSeqModule() {
+        tables = Arrays.asList(
+            new CassandraTable(CassandraMessageUidTable.TABLE_NAME,
+                SchemaBuilder.createTable(CassandraMessageUidTable.TABLE_NAME)
+                    .ifNotExists()
+                    .addPartitionKey(CassandraMessageUidTable.MAILBOX_ID, timeuuid())
+                    .addColumn(CassandraMessageUidTable.NEXT_UID, bigint())),
+            new CassandraTable(CassandraMessageModseqTable.TABLE_NAME,
+                SchemaBuilder.createTable(CassandraMessageModseqTable.TABLE_NAME)
+                    .ifNotExists()
+                    .addPartitionKey(CassandraMessageModseqTable.MAILBOX_ID, timeuuid())
+                    .addColumn(CassandraMessageModseqTable.NEXT_MODSEQ, bigint())));
+        index = Collections.emptyList();
+        types = Collections.emptyList();
+    }
+
+    @Override
+    public List<CassandraTable> moduleTables() {
+        return tables;
+    }
+
+    @Override
+    public List<CassandraIndex> moduleIndex() {
+        return index;
+    }
+
+    @Override
+    public List<CassandraType> moduleTypes() {
+        return types;
+    }
+}
